[Refactor] Notification Domain circular dependency
