+++ sseudam-core/core-application/src/main/kotlin/com/sseudam/notification/discord/DiscordClient.kt
@@ -2,7 +2,7 @@ package com.sseudam.notification.discord
 
 import com.sseudam.notification.dto.SendMessageReportDto
 import com.sseudam.notification.dto.SendMessageSuggestionDto
-import com.sseudam.user.dto.SendMessageUserProfileDto
+import com.sseudam.notification.dto.SendMessageUserProfileDto
 
 interface DiscordClient {
     fun sendCreateUserMessage(messageUserProfileDto: SendMessageUserProfileDto)
