+++ sseudam-core/core-application/src/main/kotlin/com/sseudam/visit/event/VisitedEventHandler.kt
@@ -1,14 +1,8 @@
 package com.sseudam.visit.event
 
-import com.sseudam.common.GeoConverter
-import com.sseudam.common.GeoJson
 import com.sseudam.notification.NotificationType
 import com.sseudam.notification.dto.NotificationMessages
-import com.sseudam.notification.dto.SendNotificationMessage
-import com.sseudam.notification.fcm.FcmSender
 import com.sseudam.pet.event.UserPetContextEvent
-import com.sseudam.suggestion.SuggestionService
-import com.sseudam.suggestion.SuggestionStatus
 import com.sseudam.support.extension.logger
 import com.sseudam.user.UserService
 import org.springframework.context.ApplicationEventPublisher
@@ -17,10 +11,7 @@ import org.springframework.stereotype.Component
 
 @Component
 class VisitedEventHandler(
-    private val suggestionService: SuggestionService,
     private val userService: UserService,
-    private val fcmSender: FcmSender,
-    private val geoConverter: GeoConverter,
     private val applicationEventPublisher: ApplicationEventPublisher,
 ) {
     companion object {
@@ -44,22 +35,17 @@ class VisitedEventHandler(
     )
     fun handleSendVisitNotification(event: SpotVisitedEvent) {
         try {
-            val suggestion =
-                suggestionService.findSpotSuggestionByPointAndStatus(
-                    geoConverter.geoJsonPointToJtsPoint(event.spot.point as GeoJson.Point),
-                    status = SuggestionStatus.APPROVE,
-                ) ?: return
-            val profile = userService.getProfile(suggestion.userId) ?: return
-            fcmSender.send(
-                sendNotificationMessage =
-                    SendNotificationMessage(
-                        userId = suggestion.userId,
-                        title = NotificationMessages.DEFAULT_TITLE,
-                        body = NotificationMessages.anonymousVisitedSpotContents(profile.nickname),
-                        destination = "MyPageView",
-                    ),
-                type = NotificationType.ANONYMOUS_VISITED_SPOT,
-                parameterValue = event.spot.id.toString(),
+            val suggesterId = event.spot.suggesterId ?: return
+            val profile = userService.getProfile(suggesterId) ?: return
+            applicationEventPublisher.publishEvent(
+                VisitedSpotNotificationRequestedEvent(
+                    suggesterId = suggesterId,
+                    title = NotificationMessages.DEFAULT_TITLE,
+                    body = NotificationMessages.anonymousVisitedSpotContents(profile.nickname),
+                    destination = "MyPageView",
+                    notificationType = NotificationType.ANONYMOUS_VISITED_SPOT.name,
+                    parameterValue = event.spot.id.toString(),
+                ),
             )
         } catch (e: Exception) {
             log.warn(e) { "Failed to send visit notification" }
