+++ sseudam-core/core-application/src/main/kotlin/com/sseudam/visit/event/VisitedSpotNotificationRequestedEvent.kt
@@ -1,0 +1,10 @@
+package com.sseudam.visit.event
+
+data class VisitedSpotNotificationRequestedEvent(
+    val suggesterId: Long?,
+    val title: String,
+    val body: String,
+    val destination: String,
+    val notificationType: String,
+    val parameterValue: String,
+)
